+++ src/main/core/math/Rectangle.java
@@ -12,8 +12,8 @@ public final class Rectangle {
         return x <= px && x + width >= px && y <= py && y + height >= py;
     }
 
-    public static boolean contains(int x, int y, int width, int height,
-                                   int px, int py) {
+    public static boolean contains(float x, float y, float width, float height,
+                                   float px, float py) {
         return x <= px && x + width >= px && y <= py && y + height >= py;
     }
